refactor(server): extract permission + review route domain (#591)
Move api_permission_request, api_permission_respond, api_review,
api_review_answer plus the _announce_permission_request and _live_prompt
helpers into agentwire/routes/permission.py (PermissionRoutesMixin +
register_permission_routes). Review is grouped with permission because
_live_prompt is used only by api_review.

Preserves the load-bearing aiohttp ordering (.../respond before the bare
.../permission/{name}). Part of epic #560.

Built by dotdev.dev
